+++ CHANGELOG.md
@@ -4,25 +4,50 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-https://github.com/nwn-dotnet/Anvil/compare/v8193.34.6...HEAD
+https://github.com/nwn-dotnet/Anvil/compare/v8193.34.8...HEAD
 
 ### Added
 - NwCreature - Added setters for base armor & shield arcane spell failure
+- `NuiWindowToken` - small structure with helper methods for controlling NuiWindow instances. Can be created from `NwPlayer.TryCreateNuiWindow`
 
 ### Package Updates
 - N/A
 
 ### Changed
-- N/A
+- `NwCreature.Position` will no-longer block the action queue from being modified.
+  - This fixes inconsistencies with actions queued after setting the position, but the new position can sometimes not apply for players using drive mode.
+  - For players, it is recommended to immobilize them before setting their position.
 
 ### Deprecated
-- N/A
+- `NwPlayer.TryCreateNuiWindow` - use the `NuiWindowToken` overload instead.
+- `NwPlayer.CreateNuiWindow` - use `TryCreateNuiWindow` instead.
+- `ModuleEvents.OnNuiEvent.WindowToken` - use `ModuleEvents.OnNuiEvent.Token` instead.
 
 ### Removed
 - N/A
 
 ### Fixed
-- N/A
+- Fixed `NwGameObject.Location` setter re-triggering Area Enter events.
+
+## 8193.34.8
+https://github.com/nwn-dotnet/Anvil/compare/v8193.34.7...v8193.34.8
+
+### Deprecated
+- `NwObject.Destroy(float delay)` - Use the non-delay overload instead, with the `SchedulerService`.
+
+### Fixed
+- Fixed `NwEncounter.Destroy()` not destroying the encounter object.
+- Fixed `PlaceableEvents.OnPhysicalAttacked` event always returning no data.
+
+## 8193.34.7
+https://github.com/nwn-dotnet/Anvil/compare/v8193.34.6...v8193.34.7
+
+### Package Updates
+- NWNX 6a552d9 -> 3227d60
+
+### Fixed
+- Fixed an edge case issue that caused some servers to enter an infinite crash loop when shutting down.
+- Fixed `NWNX_CORE_SHUTDOWN_SCRIPT` throwing a NRE during shutdown.
 
 ## 8193.34.6
 https://github.com/nwn-dotnet/Anvil/compare/v8193.34.5...v8193.34.6